Coverage for todo card
